requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
django-environ==0.4.5
django-heroku==0.3.1
django-localflavor==2.2
djangorestframework==3.10.3
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
gunicorn==19.9.0
oauth2client==4.1.3
psycopg2-binary==2.8.4
pip-check==2.5.2
pur==5.2.2
uritemplate==3.0.0
